Add daily and hourly forecasts to the current weather screen, save tile layout in Preferences DataStore, make minor changes to the weather screen

+++ app/src/main/java/com/example/weatherapp/core/domain/model/BriefWeather.kt
@@ -1,7 +1,7 @@
 package com.example.weatherapp.core.domain.model
 
 data class BriefWeather(
-    val temperature: Double,
+    val temperature: Int,
     val icon: String,
     val description: String,
 )
