+++ app/src/main/java/com/example/weatherapp/weather/presentation/model/WeatherHeaderInfo.kt
@@ -1,8 +1,8 @@
 package com.example.weatherapp.weather.presentation.model
 
 data class WeatherHeaderInfo(
-    val temperature: Double,
-    val feelsLike: Double,
+    val temperature: Int,
+    val feelsLike: Int,
     val description: String,
     val icon: String,
 )
